requirements.txt
serpapi
APScheduler
python-dotenv
rich
jinja2
